added tests for testing left and right hand articulated gestures
